Remove macro definition 'null'
The macro 'null' was defined to be '{0}' a zeroed struct. It was only
used to initialize pointer variables, in which case it is no different
from 0 or NULL. Removing the definition and usage of 'null' makes it
easier to use omr alongside code that defines or uses 'null' in a
different manner.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>